cmake: enable -Wall -Werror -Wextra -march=native
I had thought these were already enabled but apparently they were not.
Lots of fixes to repair all the mistakes found by these warnings.

The correct way to add these options in cmake seems to be
the 'add_compile_options' directive.

That should also apply then to any subdirecties below where it is set.

[1] https://cmake.org/cmake/help/latest/command/add_compile_options.html#command

+++ ast/ast.h
@@ -39,16 +39,16 @@ struct Namespace {
 
 	// enum declarations
 	struct EnumDecl** enums;
-	uint16_t count_enums;
+	size_t count_enums;
 	size_t capacity_enums;
 
 	//structs must be declared before the subroutines
 	struct StructDecl** structs;
-	uint16_t count_structs;
+	size_t count_structs;
 	size_t capacity_structs;
 
 	struct Method** methods;
-	uint16_t count_methods;
+	size_t count_methods;
 	size_t capacity_methods;
 };
